HandwerkerRechner: Hoch/Mittel-Fixes - Message-Overlay, Ad-Opt-in, Services, decimal
- MessageRequested hatte app-weit keinen Abonnenten (alle Meldungen verpufften):
  Inline-Message-Overlay in MainView + ShowMessage-Routing aller Call-Sites
- Ad-Opt-in statt erzwungenem Rewarded nach jeder 3. Berechnung (AdMob-Policy):
  Dialog "Video ansehen -> Banner 10 Berechnungen ausblenden", echte Belohnung,
  expliziter AdConfig-Arm calculation_ad, 4 neue RESX-Keys in 6 Sprachen
- Back-Button-Kaskade vervollstaendigt (Message-/Ad-Dialog, Projekt-Dialoge,
  Template-Apply, Quote-GoBack mit IsEditing) + UpdateFavorites bei Sprachwechsel
- Activity-Leak: ExitHintRequested-Handler wird in OnDestroy abgemeldet
- LoadingPipeline: MainViewModel-Graph auf UI-Thread (Brush-Affinitaet),
  App-Fehlerpfad mit DataContext-Recovery statt leerer UI, Splash-Version aus Assembly
- Services: SaveFailed-Events + Semaphore-Hygiene in ProjectTemplate-/
  MaterialPriceService, .bak-Sicherung korrupter JSON vor Reset
- Geldbetraege auf decimal (Quote/QuoteItem/MaterialPrice, AwayFromZero-Rundung,
  CustomPrice-Sentinel -1 -> nullable), QuoteViewModel-Totals dedupliziert,
  Angebotsnummer mit lokalem Jahr
- Kleinkram: Debounce-Timer-Nullen (Grout/Metal), Metal-Dimension2-Validierung,
  Aufmass-Negativ-Guards fuer alle 6 Formen, Banner-Margins History/Settings,
  x:DataType am Favoriten-Template, tote englische Defaults entfernt
- Tests: 81/81 gruen (13 neue: decimal-Rundung, JSON-Roundtrip, Sentinel-Migration)

+++ src/Apps/HandwerkerRechner/HandwerkerRechner.Shared/ViewModels/Floor/ConcreteCalculatorViewModel.cs
@@ -76,7 +76,7 @@ public sealed partial class ConcreteCalculatorViewModel : ViewModelBase, IDispos
         _unitConverter.UnitSystemChanged += OnUnitSystemChanged;
 
         // Standard-Materialpreis laden
-        PricePerBag = _priceService.GetPrice("concrete_sack_25kg")?.EffectivePrice ?? 0;
+        PricePerBag = (double)(_priceService.GetPrice("concrete_sack_25kg")?.EffectivePrice ?? 0);
     }
 
     /// <summary>
